fix: include DM recipients from 'to' array in mentions for incoming remote DMs
When receiving DMs from remote servers, the local user (recipient) was not
being added to the mentions array because some ActivityPub servers only
include recipients in the 'to' field without corresponding Mention tags.

This caused remote DMs to not appear in the conversation list since the SQL
query relies on the mentions array to find DMs where the user is a recipient.

Changes:
- Modified extractMentions to also process dmRecipients parameter
- For DMs (specified visibility), pass the 'to' array to extractMentions
- Added duplicate prevention using Set to avoid adding same user twice
- Added debug logging for DM recipient resolution